# openapi-ima-pro/openapi.yaml
openapi: 3.0.3
info:
  title: New API - IMA Pro（OpenAI 兼容）
  version: 1.5.14
  description: "IMA Pro 对外接口文档（严格对齐 New API 当前实现）。\n\n版本更新（1.5.14）：\n- `/v1/videos` 新增 Seedance 模型支持（同 Ima-Pro 端点）：\n  - `seedance-2.0`\n  - `seedance-2.0-fast`\n  - `seedance-2.0-cn`\n  - `seedance-2.0-fast-cn`\n\
    - Asset 上传模型补充海外模型说明：\n  - `seedance-upload`（海外）\n\n版本更新（1.5.13）：\n- Asset 接口请求体新增 `model` 字段（支持按模型做权限/渠道映射）：\n  - `/v1/assets/group/create`\n  - `/v1/assets/group/list`\n  - `/v1/assets/create`\n \
    \ - `/v1/assets/list`\n- 明确计费行为：\n  - 仅 `/v1/assets/create` 会触发扣费\n  - 扣费模型与请求体 `model` 一致（未传时默认 `ima-pro-upload`）\n\n版本更新（1.5.12）：\n- 新增“音乐版权风险规避”反例示例（C14）：\n  - 建议先禁用背景音乐（`metadata.audio=false`），避免潜在版权争议\n\
    \  - 提示词约束“无字幕、无背景音乐，仅保留剧情人声/环境音效/动作音效”\n\n版本更新（1.5.11）：\n- 补充 `asset://` 素材协议示例（包含人像参考图场景）：\n  - `POST /v1/videos` 请求示例新增“文本 + 图片(公网 URL + asset://) + 视频 + 音频”案例\n  - 明确 `images` / `input_reference`\
    \ 均支持 `asset://asset-xxxx` 与 `https://` URL\n\n版本更新（1.5.10）：\n- 明确 `/v1/videos` 仅用于视频任务：\n  - 仅支持：`ima-pro`、`ima-pro-fast`\n  - 图片模型（如 `gemini-3-pro-image-preview`、`gemini-3.1-flash-image-preview`）不支持该端点\n\
    - 补充图片模型误用 `/v1/videos` 的 400 错误示例（`invalid_model_endpoint`）\n\n版本更新（1.5.9）：\n- 新增 Asset 对外端点文档：\n  - `/v1/assets/group/create|list|get|update`\n  - `/v1/assets/create|list|get|update|delete|quota`\n\
...
    \ 当前对外行为编写。\n"
servers:
- url: https://open-route.fashionlabs.cn/
- url: https://api.imarouter.com/
...
security:
- BearerAuth: []
paths:
...
  /v1/assets/group/create:
    post:
      summary: 创建素材分组
      description: '创建素材分组（不扣费）。

        鉴权：请求头必须包含 `Authorization: Bearer sk-...`。


        字段说明：

        - `name`：分组名称（必填，<=64 字符）

        - `description`：分组描述（可选，<=300 字符）

        - `group_type`：分组类型（可选；默认 `AIGC`）

        - `project_name`：项目名（可选；默认 `default`）

        - `model`：可选，资产模型名（建议显式传，避免默认模型与分组权限不匹配）

        '
      operationId: createAssetGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetGroupCreateRequest'
            examples:
              minimal:
                summary: 最小可用请求
                value:
                  name: marketing-assets
              full:
                summary: 带可选字段的请求
                value:
                  name: marketing-assets
                  description: 素材分组-市场投放
                  group_type: AIGC
                  project_name: default
                  model: ima-pro-upload-cn
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetGroupResponse'
      tags:
      - 人像
      x-apifox-folder: 人像
  /v1/assets/group/list:
    post:
      summary: 查询素材分组列表
      description: '查询分组列表（不扣费）。

        排序参数严格校验：

        - `sort_by`: `CreateTime|UpdateTime`

        - `sort_order`: `Asc|Desc`

        - `model`: 可选，建议显式传用于按模型权限过滤。

        '
      operationId: listAssetGroups
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetGroupListRequest'
            examples:
              default_page:
                summary: 默认分页查询
                value:
                  page_number: 1
                  page_size: 20
                  model: ima-pro-upload-cn
              with_filter:
                summary: 按名称过滤并按时间倒序
                value:
                  filters:
                    name: marketing
                    group_type: AIGC
                  sort_by: CreateTime
                  sort_order: Desc
                  page_number: 1
                  page_size: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetGroupListResponse'
      tags:
      - 人像
      x-apifox-folder: 人像
  /v1/assets/group/get:
    post:
      summary: 查询单个素材分组
      operationId: getAssetGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - id
              properties:
                id:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetGroupResponse'
      tags:
      - 人像
      x-apifox-folder: 人像
  /v1/assets/group/update:
    post:
      summary: 更新素材分组
      operationId: updateAssetGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - id
              properties:
                id:
                  type: string
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetIdResponse'
      tags:
      - 人像
      x-apifox-folder: 人像
  /v1/assets/create:
    post:
      summary: 上传素材（创建素材）
      description: '上传素材 URL 到素材库。

        计费说明：仅此端点扣费，计费模型取请求体 `model`（未传时默认 `ima-pro-upload`，按次）。

        默认扣费令牌：当前请求头 `Authorization` 对应的 SK。

        可选字段 `billing_token_id` 仅用于覆盖默认扣费令牌（必须是当前用户且启用）。

        `asset_type` 表示素材类型，默认值 `Image`，支持 `Image|Video|Audio`（不传则按 `Image`）。

        `asset_type=Audio` 时，音频时长由上游服务校验，需在 `1.8s` 到 `15.2s` 闭区间内，超限会报：
        `InvalidParameter.DurationTooLong: Duration must be between 1.8s and 15.2s`。

        '
      operationId: createAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetCreateRequest'
            examples:
              video_create_success:
                summary: 视频素材上传示例（成功）
                value:
                  group_id: group-20260403210537-6v56x
                  url: https://ws.esxscloud.com/aiagent/src/d/20260407/in/1a2011e98a41847d70a49b570995d379.mp4
                  name: verify-overseas-video-check
                  asset_type: Video
                  model: ima-pro-upload-cn
              image_create_success:
                summary: 图片素材上传示例（成功）
                value:
                  group_id: group-20260403210537-6v56x
                  url: https://dxii5p4krscl5.cloudfront.net/resource/image/dd40f8e8-9296-4daf-9277-4a842d79fbee.jpg
                  name: verify-overseas-image-check
                  asset_type: Image
                  model: ima-pro-upload-cn
              audio_create_success:
                summary: 音频素材上传示例（成功）
                value:
                  group_id: group-20260403210537-6v56x
                  url: https://ima-ga.esxscloud.com/drama_pro/privite/2026/04/02/1775124683350_1930188605918449664_35ae6ccc808744e5bcb2fc2354bf9c1c.mp3
                  name: verify-overseas-audio-check
                  asset_type: Audio
                  model: ima-pro-upload-cn
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetIdResponse'
              examples:
                create_success:
                  value:
                    success: true
                    message: ""
                    data:
                      Id: asset-20260409110407-jpbr5
        '400':
          description: '参数错误。
            常见错误示例：
            - `InvalidParameter.DurationTooLong: Duration must be between 1.8s and 15.2s`
            - `doubao asset request failed: InvalidParameter.DownloadFailed`
            '
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
              examples:
                invalid_audio_duration:
                  value:
                    success: false
                    message: "doubao asset request failed: InvalidParameter.DurationTooLong: Duration must be between 1.8s and 15.2s"
      tags:
      - 人像
      x-apifox-folder: 人像
  /v1/assets/list:
    post:
      summary: 查询素材列表
      description: '查询素材列表（不扣费）。

        排序参数严格校验：

        - `sort_by`: `CreateTime|UpdateTime|GroupId`

        - `sort_order`: `Asc|Desc`

        - `model`: 可选，建议显式传用于按模型权限过滤。

        '
      operationId: listAssets
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetListRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetListResponse'
      tags:
      - 人像
      x-apifox-folder: 人像
  /v1/assets/get:
    post:
      summary: 查询单个素材
      operationId: getAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - id
              properties:
                id:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetGetResponse'
      tags:
      - 人像
      x-apifox-folder: 人像
  /v1/assets/update:
    post:
      summary: 更新素材名称
      operationId: updateAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - id
              - name
              properties:
                id:
                  type: string
                name:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetIdResponse'
      tags:
      - 人像
      x-apifox-folder: 人像
  /v1/assets/quota:
    post:
      summary: 查询素材配额信息
      operationId: getAssetQuota
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetQuotaResponse'
      tags:
      - 人像
      x-apifox-folder: 人像
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sk-xxxx
      description: '将 SK 放在请求头：

        `Authorization: Bearer sk-xxxx`

        '
  schemas:
    AssetGroupCreateRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          maxLength: 64
          description: 分组名称（用户侧展示名）。
          example: marketing-assets
        description:
          type: string
          maxLength: 300
          description: 分组描述（可选）。
          example: 素材分组-市场投放
        group_type:
          type: string
          default: AIGC
          description: 分组类型。当前默认使用 `AIGC`。
          example: AIGC
        project_name:
          type: string
          default: default
          description: 项目命名空间。未传时默认 `default`。
          example: default
        model:
          type: string
          enum:
          - ima-pro-upload
          - ima-pro-upload-cn
          - seedance-upload
          description: '可选，资产模型名。建议显式传，避免默认模型与当前分组权限不匹配。

            - `ima-pro-upload`：国内人像（默认）

            - `ima-pro-upload-cn`：国内人像

            - `seedance-upload`：海外人像

            '
          example: ima-pro-upload-cn
    AssetGroupListRequest:
      type: object
      properties:
        filters:
          type: object
          properties:
            name:
              type: string
            group_ids:
              type: array
              items:
                type: string
            group_type:
              type: string
              default: AIGC
              description: 分组类型过滤，通常为 `AIGC`。
        page_number:
          type: integer
          minimum: 1
          default: 1
          description: 页码（从 1 开始）。
        page_size:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
          description: 每页数量（1-100）。
        sort_by:
          type: string
          enum:
          - CreateTime
          - UpdateTime
          description: 排序字段（严格区分大小写）。
        sort_order:
          type: string
          enum:
          - Asc
          - Desc
          description: 排序方向（严格区分大小写）。
        project_name:
          type: string
          default: default
          description: 项目命名空间。未传按默认项目处理。
        model:
          type: string
          enum:
          - ima-pro-upload
          - ima-pro-upload-cn
          - seedance-upload
          description: '可选，资产模型名。建议显式传，避免默认模型与当前分组权限不匹配。

            - `ima-pro-upload`：国内人像（默认）

            - `ima-pro-upload-cn`：国内人像

            - `seedance-upload`：海外人像

            '
          example: ima-pro-upload-cn
    AssetCreateRequest:
      type: object
      required:
      - group_id
      - url
      properties:
        group_id:
          type: string
          description: 分组 ID（可用 `/v1/assets/group/list` 获取）。
          example: grp_123456
        url:
          type: string
          format: uri
          description: 公开可访问的素材 URL。
          example: https://example.com/image.png
        name:
          type: string
          maxLength: 64
          description: 素材名称（可选）。
          example: cover-image
        asset_type:
          type: string
          default: Image
          enum:
          - Image
          - Video
          - Audio
          description: '素材类型，默认 `Image`。

            当 `asset_type=Audio` 时，音频时长必须在 `1.8s` 到 `15.2s` 之间（闭区间），否则返回 `InvalidParameter.DurationTooLong`。'
          example: Image
        project_name:
          type: string
          default: default
          description: 项目命名空间。未传时默认 `default`。
          example: default
        billing_token_id:
          type: integer
          description: '可选，指定本次上传扣费使用的 token id。

            不传时：默认使用当前请求头 `Authorization` 对应 SK 的 token。

            约束：必须属于当前用户、状态为启用、并且余额足够。

            '
          example: 3
        model:
          type: string
          enum:
          - ima-pro-upload
          - ima-pro-upload-cn
          - seedance-upload
          description: '可选，资产模型名。

            计费时优先使用该模型价格；未传时默认 `ima-pro-upload`。

            - `ima-pro-upload`：国内人像（默认）

            - `ima-pro-upload-cn`：国内人像

            - `seedance-upload`：海外人像

            '
          example: ima-pro-upload-cn
    AssetListRequest:
      type: object
      properties:
        filters:
          type: object
          properties:
            group_ids:
              type: array
              items:
                type: string
            group_type:
              type: string
              default: AIGC
              description: 分组类型过滤，通常为 `AIGC`。
            statuses:
              type: array
              items:
                type: string
                enum:
                - Active
                - Failed
                - Processing
            name:
              type: string
        page_number:
          type: integer
          minimum: 1
          default: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        sort_by:
          type: string
          enum:
          - CreateTime
          - UpdateTime
          - GroupId
          description: 排序字段（严格区分大小写）。
        model:
          type: string
          enum:
          - ima-pro-upload
          - ima-pro-upload-cn
          - seedance-upload
          description: '可选，资产模型名。建议显式传，避免默认模型与当前分组权限不匹配。

            - `ima-pro-upload`：国内人像（默认）

            - `ima-pro-upload-cn`：国内人像

            - `seedance-upload`：海外人像

            '
          example: ima-pro-upload-cn
        sort_order:
          type: string
          enum:
          - Asc
          - Desc
          description: 排序方向（严格区分大小写）。
        project_name:
          type: string
          default: default
          description: 项目命名空间。未传时默认 `default`。
    AssetItem:
      type: object
      properties:
        Id:
          type: string
        Name:
          type: string
        URL:
          type: string
        GroupId:
          type: string
        AssetType:
          type: string
        Status:
          type: string
        ProjectName:
          type: string
        CreateTime:
          type: string
        UpdateTime:
          type: string
    AssetGroupItem:
      type: object
      properties:
        Id:
          type: string
        Name:
          type: string
        Title:
          type: string
        Description:
          type: string
        GroupType:
          type: string
        ProjectName:
          type: string
        CreateTime:
          type: string
        UpdateTime:
          type: string
    AssetIdResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            Id:
              type: string
    AssetGroupResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/AssetGroupItem'
    AssetGetResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/AssetItem'
    AssetGroupListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            Items:
              type: array
              items:
                $ref: '#/components/schemas/AssetGroupItem'
            TotalCount:
              type: integer
            PageNumber:
              type: integer
            PageSize:
              type: integer
    AssetListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            Items:
              type: array
              items:
                $ref: '#/components/schemas/AssetItem'
            TotalCount:
              type: integer
            PageNumber:
              type: integer
            PageSize:
              type: integer
    AssetQuotaResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            enabled:
              type: boolean
            max_count_per_user:
              type: integer
            current_count:
              type: integer
...